HOHQMesh.jl API

HOHQMesh.examples_dir — Method

examples_dir()

Return the path to the directory with some example mesh setups.

HOHQMesh.generate_mesh — Method

generate_mesh(control_file;
              output_directory="out",
              mesh_filename=nothing, plot_filename=nothing, stats_filename=nothing)

Generate a mesh based on the control_file with the HOHQMesh mesh generator and store resulting files in output_directory.

You can set the mesh filename, the plot filename, and the statistics filename using the keyword arguments mesh_filename, plot_filename, and stats_filename, respectively. If set to nothing, the filenames for the mesh file, plot file, and statistics file are generated automatically from the control file name. For example, path/to/ControlFile.control will result in output files ControlFile.mesh, ControlFile.tec, and ControlFile.txt.

You can activate verbose output from HOHQMesh that prints additional messages and debugging mesh information with the keyword argument verbose.

This function returns the output to stdout of the HOHQMesh binary when generating the mesh.