HOHQMesh.jl API

HOHQMesh.add! — Method

add!(proj::Project, crv::Dict{String,Any}, boundaryName::String)

Add a curve to the inner boundary named boundaryName.

HOHQMesh.add! — Method

add!(proj::Project, obj::Dict{String,Any})

Add a curve to the outer boundary or a refinement reion to the project

HOHQMesh.addBackgroundGrid! — Method

addBackgroundGrid!(proj::Project, x0::Array{Float64}, dx::Array{Float64}, N::Array{Int})

Add the background grid block using the left corner, x0, the grid size dx, and the number of intervals in each direction. Use this when there is no outer boundary defined in the model. This version mimics HOHQMesh's backgroundGrid block, but the version

addBackgroundGrid!(proj::Project, box::Array{Float64},  N::Array{Int} )

is a lot easier to use.

TODO: Change HOHQMesh and delete this way to specify the domain and use the bounding box one instead.

HOHQMesh.addBackgroundGrid! — Method

addBackgroundGrid!(proj::Project, box::Array{Float64},  N::Array{Int} )

Add the background grid block with bounding box = [TOP, LEFT, BOTTOM, RIGHT] and the number of intervals in each diredction. Use this when there is no outer boundary defined in the model.

HOHQMesh.addBackgroundGrid! — Method

addBackgroundGrid!(proj::Project, bgSize::Array{Float64})

Add the background grid block with the grid size to be a 3-vector. Use this when there is an outer boundary defined in the model.

HOHQMesh.addBackgroundGrid! — Method

addBackgroundGrid!(proj::Project, dict::Dict{String,Any})

Used only for undo/redo.

HOHQMesh.addCurveToInnerBoundary! — Method

addCurveToInnerBoundary!(proj::Project, crv::Dict{String,Any}, boundaryName::String)

Add a curve to the inner boundary with name boundaryName. If an inner boundary of that name does not exist, one is created.

HOHQMesh.addCurveToOuterBoundary! — Method

addCurveToOuterBoundary!(proj::Project, crv::Dict{String,Any})

Add a curve to the outer boundary. The curves must be added in order counter-clockwise

HOHQMesh.addInnerBoundaryWithName! — Method

addInnerBoundaryWithName!(proj::Project,name::String)

Create a new empty inner boundary with the given name.

HOHQMesh.addObserver — Method

addObserver(observer::Any, note::String, f::Any)

f is the function to be executed (called) when a notification of name note is given.

The function called upon notification must have the signature f(observer, sender, args...)

HOHQMesh.addOuterBoundary! — Method

addOuterBoundary!(proj::Project, outerBoundary::Dict{String,Any})

Add an empty outer boundary to the project. There can be only one. This function is only used as part of an undo operation removing the outer boundary.

HOHQMesh.addRefinementRegion! — Method

addRefinementRegion!(proj::Project,r::Dict{String,Any})

Add the refinement region to the project

HOHQMesh.addRefinementRegionPoints! — Method

addRefinementRegionPoints!(proj::Project, r::Dict{String,Any})

Compute and add to the project the plotting points for the refinement region

HOHQMesh.addRunParameters! — Function

addRunParameters!(proj::Project,
                  plotFormat::String     = "skeleton",
                  meshFileFormat::String = "ISM-V2",
                  polynomialOrder::Int   = 5)

Add a RUN_PARAMETERS block and set all the parameters in one call.

HOHQMesh.addSpringSmoother! — Function

addSpringSmoother!(status::String, type::String, nIterations::Int)

Status is either ON or OFF Type is either LinearSpring or LinearAndCrossbarSpring

HOHQMesh.bboxUnion — Method

bboxUnion(box1::Array{Float64}, box2::Array{Float64})

Returns the union of two bounding boxes

HOHQMesh.curveBounds — Method

curveBounds(crvPoints::Array{Float64,2})

Find the bounds of a single curve, discretized as an array

HOHQMesh.emptyBounds — Method

emptyBounds()

Returns an array that will always be ignored when unioned with another bounding box.

HOHQMesh.examples_dir — Method

examples_dir()

Return the path to the directory with some example mesh setups.

HOHQMesh.extract_mesh_file_format — Method

extract_mesh_file_format(control_file)

Return a string with the desired output format of the HOHQMesh generated mesh file. This information is given within the RUN_PARAMETERS of the CONTROL_INPUT block of the control file. See the HOHQMesh documentation for details.

HOHQMesh.generate_mesh — Method

generate_mesh(control_file;
              output_directory="out",
              mesh_filename=nothing, plot_filename=nothing, stats_filename=nothing,
              verbose=false, subdivision_maximum=8)

Generate a mesh based on the control_file with the HOHQMesh mesh generator and store resulting files in output_directory.

You can set the mesh filename, the plot filename, and the statistics filename using the keyword arguments mesh_filename, plot_filename, and stats_filename, respectively. If set to nothing, the filenames for the mesh file, plot file, and statistics file are generated automatically from the control file name. For example, path/to/ControlFile.control will result in output files ControlFile.mesh, ControlFile.tec, and ControlFile.txt.

You can activate verbose output from HOHQMesh that prints additional messages and debugging mesh information with the keyword argument verbose.

To override the maximum number of allowable subdivisions in the quad tree during meshing adjust the value of subdivision_maximum. The default value of subdivision_maximum is 8, meaning that elements can be up to a factor of 2^8 smaller than the background grid. Note, think before doing this! It could be adjusting the boundary curves, background grid size, adding local refinement regions, or some combination may remove the need to adjust the subdivision depth.

This function returns the output to stdout of the HOHQMesh binary when generating the mesh.

HOHQMesh.generate_mesh — Method

generate_mesh(proj::Project; verbose=false, subdivision_maximum=8)

Generate a mesh from the information stored in a Project created using the interactive mesh functionality. First a check is made if a background grid exists and all inner/outer boundary curves are valid.

This function will then make a HOHQMesh control file from the control dictionary proj.controlDict and use it to call the wrapper function that interfaces with HOHQMesh. The resulting mesh and control files will be saved to proj.projectDirectory. Also, if there is an active plot of the mesh project it will update to display the generated mesh.

With the optional argument verbose one can activate verbose output from HOHQMesh that prints additional messages and debugging mesh information.

To override the maximum number of allowable subdivisions in the quad tree during meshing adjust the value of subdivision_maximum. The default value of subdivision_maximum is 8, meaning that elements can be up to a factor of 2^8 smaller than the background grid. Note, think before doing this! It could be adjusting the boundary curves, background grid size, adding local refinement regions, or some combination may remove the need to adjust the subdivision depth.

This function returns the output to stdout of the HOHQMesh binary when generating the mesh.

HOHQMesh.getAllInnerBoundaries — Method

getAllInnerBoundaries(proj::Project)

Returns an array of the inner boundaries

HOHQMesh.getAllRefinementRegions — Method

getAllRefinementRegions(proj::Project)

Get the list of refinement regions.

HOHQMesh.getArcCenter — Method

getArcCenter(crv::Dict{String,Any})

Get the center of a circular or elliptic arc as an array

HOHQMesh.getArcEndAngle — Method

getArcEndAngle(arc::Dict{String,Any})

Returns the end angle of a CIRCULAR_ARC or ELLIPTIC_ARC.

HOHQMesh.getArcRadius — Function

getArcRadius(arc::Dict{String,Any}, key::String = "radius")

Returns the current radius of an arc. Default behavior expects a CIRCULAR_ARC with key = "radius". For an ELLIPTIC_ARC use either key = "xRadius" or key = "yRadius".

HOHQMesh.getArcRotation — Method

getArcRotation(arc::Dict{String,Any})

Returns the current rotation angle of an ELLIPTIC_ARC.

HOHQMesh.getArcStartAngle — Method

getArcStartAngle(arc::Dict{String,Any})

Returns the start angle of a CIRCULAR_ARC or ELLIPTIC_ARC.

HOHQMesh.getArcUnits — Method

getArcUnits(crv::Dict{String,Any})

Get the units for the start and end angles of a circular arc curve.

HOHQMesh.getBackgroundGridLowerLeft — Method

getBackgroundGridLowerLeft(proj::Project)

Returns the [x,y] of the lower left point of the background grid.

HOHQMesh.getBackgroundGridSize — Method

getBackgroundGridSize(proj::Project)

Returns the background grid size array.

HOHQMesh.getBackgroundGridSteps — Method

getBackgroundGridSteps(proj::Project)

Returns the [Nx,Ny,Nz] for the background grid.

HOHQMesh.getCurve — Method

getCurve(proj::Project, curveName::String, boundaryName::String)

Get the curve named curveName from the inner boundary named boundaryName

HOHQMesh.getCurve — Method

getCurve(proj::Project, curveName::String)

Get the curve with name curveName from the outer boundary.

HOHQMesh.getCurveName — Method

getCurveName(crv::Dict{String,Any})

Returns the "name" of a curve stored in crv.

HOHQMesh.getCurveType — Method

getCurveType(crv::Dic{String,Any})

Get the type of the curve stored in crv. Generic for any curve dictionary END_POINTSLINE_CURVE, PARAMETRIC_EQUATION_CURVE, SPLINE_CURVE, CIRCULAR_ARC, or ELLIPTIC_ARC as a string.

HOHQMesh.getEndPoint — Method

getEndPoint(crv::Dict{String,Any})

Get the end point for a line curve as an array.

HOHQMesh.getInnerBoundary — Method

getInnerBoundary(proj::Project, name::String)

Get the chain of curves from the inner boundary with name name.

HOHQMesh.getInnerBoundaryChainConnect — Method

getInnerBoundaryChainConnect(proj::Project, chainName::String)

Returns the current connect setting for the boundary curve optimization of an inner boundary CHAIN with chainName.

HOHQMesh.getInnerBoundaryChainContinuity — Method

getInnerBoundaryChainContinuity(proj::Project, chainName::String)

Returns the current continuity (number of continuous derivatives) for the boundary curve optimization of an inner boundary CHAIN with chainName.

HOHQMesh.getInnerBoundaryChainOptimizeStatus — Method

getInnerBoundaryChainOptimizeStatus(proj::Project, chainName::String)

Returns the current type of the boundary curve optimization of an inner boundary CHAIN with chainName.

HOHQMesh.getInnerBoundaryChainTolerance — Method

getInnerBoundaryChainTolerance(proj::Project, chainName::String)

Returns the current tolerance for the boundary curve optimization of an inner boundary CHAIN with chainName.

HOHQMesh.getInnerBoundaryChainWithName — Method

getInnerBoundaryChainWithName(proj::Project, name::String)

Get the inner boundary CHAIN with the given name. If one does not exist, it is created.

HOHQMesh.getMeshFileFormat — Method

getMeshFileFormat(proj::Project)

Returns the format in which the mesh will be written.

HOHQMesh.getName — Method

getName(proj::Project)

Returns the filename to be used by the mesh, plot, control, and stats files.

HOHQMesh.getOuterBoundaryChainList — Method

getOuterBoundaryChainList(proj::Project)

Get the array of outer boundary curves.

HOHQMesh.getOuterBoundaryConnect — Method

getOuterBoundaryConnect(proj::Project)

Returns the current connect setting for the boundary optimization of the outer boundary chain.

HOHQMesh.getOuterBoundaryContinuity — Method

getOuterBoundaryContinuity(proj::Project)

Returns the current continuity (number of continuous derivatives) for the boundary optimization of the outer boundary chain.

HOHQMesh.getOuterBoundaryCurveWithName — Method

getOuterBoundaryCurveWithName(proj::Project, name::String)

Return the curve dictionary crv with name from the OUTER_BOUNDARY curve chain.

HOHQMesh.getOuterBoundaryOptimizeStatus — Method

getOuterBoundaryOptimizeStatus(proj::Project)

Returns the current type for the boundary curve optimization for the outer boundary chain.

HOHQMesh.getOuterBoundaryTolerance — Method

getOuterBoundaryTolerance(proj::Project)

Returns the current tolerance for the boundary optimization of the outer boundary chain.

HOHQMesh.getPlotFileFormat — Method

getPlotFileFormat(proj::Project)

Returns the plot file format.

HOHQMesh.getPolynomialOrder — Method

getPolynomialOrder(proj::Project)

Returns the polynomial order for boundary curves in the mesh file.

HOHQMesh.getRefinementEnd — Method

getRefinementEnd(r::Dict{String,Any})

Return Array{Float64} of the end location of the refinement line

HOHQMesh.getRefinementGridSize — Method

getRefinementGridSize(r::Dict{String,Any})

Returns the grid size,h, as Float64. r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementLocation — Method

getRefinementLocation(r::Dict{String,Any})

Return Array{Float64} of the location of the refinement center.r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementName — Method

getRefinementName(r::Dict{String,Any})

Return name of the refinement. r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementRegion — Method

getRefinementRegion(proj::Project, index)

Get the refinement region with index, index from the project. Returns nothing if there is none. The return value is a dictionary that represents the refinement region.

HOHQMesh.getRefinementRegion — Method

(i,r) = getRefinementRegion(project, name)

Get the refinement region with the given name and its location in the list of refinement regions.

HOHQMesh.getRefinementRegionCenter — Method

getRefinementRegionCenter(r::Dict{String,Any})

Get, or compute, the center of the given refinement region.

HOHQMesh.getRefinementStart — Method

getRefinementStart(r::Dict{String,Any})

Return Array{Float64} of the start location of the refinement line. r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementType — Method

getRefinementType(r::Dict{String,Any})

Return the type of refinement, either "smooth" or "sharp". r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementWidth — Method

getRefinementWidth(r::Dict{String,Any})

Returns the region width,w, as Float64. r is the dictionary that represents the refinement region.

HOHQMesh.getSmoothingIterations — Method

getSmoothingIterations(proj::Project)

Get the number of iterations to smooth the mesh.

HOHQMesh.getSmoothingStatus — Method

getSmoothingStatus(proj::Project)

Returns whether the smoother will be "ON" or "OFF"

HOHQMesh.getSmoothingType — Method

getSmoothingType(proj::Project)

Returns either "LinearSpring" or "LinearAndCrossbarSpring"

HOHQMesh.getSplineNKnots — Method

getSplineNKnots(spline::Dict{String,Any})

Return the number of points (or knots) for the spline.

HOHQMesh.getSplinePoints — Method

getSplinePoints(spline::Dict{String,Any})

Return the parametric spline point data from the spline curve dictionary.

HOHQMesh.getStartPoint — Method

getStartPoint(crv::Dict{String,Any})

Get the start point for a line curve as an array

HOHQMesh.getXEqn — Method

getXEqn(crv::Dict{String,Any})

Return the expression for the x(t) equation of a PARAMETRIC_EQUATION_CURVE.

HOHQMesh.getYEqn — Method

getYEqn(crv::Dict{String,Any})

Return the expression for the y(t) equation of a PARAMETRIC_EQUATION_CURVE.

HOHQMesh.getZEqn — Method

getZEqn(crv::Dict{String,Any})

Return the expression for the z(t) equation of a PARAMETRIC_EQUATION_CURVE.

HOHQMesh.hasBackgroundGrid — Method

hasBackgroundGrid(proj::Project)

Tests to see if the project has a backgroundGrid dictionary defined.

HOHQMesh.innerBoundaryIndices — Method

innerBoundaryIndices(proj::Project, curveName::String)

Returns (curveIndex,chainIndex) for the location of the curve named curveName in it's inner boundary chain.

HOHQMesh.insertInnerBoundaryAtIndex! — Method

insertInnerBoundaryAtIndex!(proj::Project, chainName::String, index::Int, chain::??)

Insert an entire inner boundary. Primarily meant for undo operation.

HOHQMesh.insertInnerBoundaryCurveAtIndex! — Method

insertInnerBoundaryCurveAtIndex!(proj::Project, crv::Dict{String,Any},
                                 index::Int, boundaryName::String)

Insert a curve crv into an inner boundary chain boundaryName at the specified index index.

HOHQMesh.insertOuterBoundaryCurveAtIndex! — Method

insertOuterBoundaryCurveAtIndex!(proj::Project, crv::Dict{String,Any}, index::Int)

Insert a curve into the outer boundary chain at the specified index.

HOHQMesh.insertRefinementRegion! — Method

insertRefinementRegion!(proj::Project, r::Dict{String,Any}, index::Int)

Used by undo()

HOHQMesh.modelChainIsOK — Method

modelChainIsOK(chain::Vector{Dict{String, Any}}, chainName::String)

Returns true if the chain of curves is contiguous and closed; false otherwise.

HOHQMesh.modelCurvesAreOK — Method

modelCurvesAreOK(proj::Project)

Go through all curves in the model and make sure they are connected and closed. Also, remove any empty outer / inner boundary chains.

Returns true if all curves are connected and closed, false otherwise.

HOHQMesh.new — Function

new(name::String,
    xEqn::String,
    yEqn::String,
    zEqn::String = "z(t) = 0.0" )

Create a new parametric equation curve.

HOHQMesh.new — Function

new(name::String,
    center::Array{Float64},
    radius::Float64,
    startAngle::Float64,
    endAngle::Float64,
    units::String)

Create a new circular arc.

HOHQMesh.new — Function

new(name::String,
    center::Array{Float64},
    xradius::Float64,
    yradius::Float64,
    startAngle::Float64,
    endAngle::Float64,
    rotation::Float64,
    units::String)

Create a new elliptic arc.

HOHQMesh.new — Method

new(name::String,
         xStart::Array{Float64},
         xEnd::Array{Float64})

Create a new line defined by its end points.

HOHQMesh.new — Method

new(name::String, nKnots::Int, data::Matrix{Float64})

Create a spline curve from an array of knots

HOHQMesh.new — Method

new(name::String, dataFile::String)

Create a spline curve from the contents of a data file.

HOHQMesh.newCircularArcCurve — Function

newCircularArcCurve(name::String, center::Array{Float64},
    radius::Float64, startAngle::Float64, endAngle::Float64,
    units::String)

Creates and returns a new circular arc curve in the form of a Dictionary.

HOHQMesh.newEllipticArcCurve — Function

newEllipticArcCurve(name::String, center::Array{Float64},
                    xRadius::Float64, yRadius::Float64,
                    startAngle::Float64, endAngle::Float64,
                    rotation::Float64, units::String)

Creates and returns a new elliptic arc curve in the form of a Dictionary.

HOHQMesh.newEndPointsLineCurve — Method

newEndPointsLineCurve(name::String, xStart::Array{Float64},xEnd::Array[Float64])

Creates and returns a new curve defined by its end points in the form of a Dictionary

HOHQMesh.newParametricEquationCurve — Function

newParametricEquationCurve(name::String,
                           xEqn::String,
                           yEqn::String,
                           zEqn::String = "z(t) = 0.0" )

Creates and returns a new parametricEquationCurve in the form of a Dictionary

HOHQMesh.newProject — Method

newProject(name::String, folder::String)

Create a new project with the given name. That name will be used for the mesh and plot files in the specified folder.

HOHQMesh.newRefinementCenter — Method

newRefinementCenter(name, type,
                    center, meshSize,
                    width)

Create refinement center of type "smooth" or "sharp" centered at center = [x,y,z] with a mesh size meshSize spread over a radius width.

HOHQMesh.newRefinementLine — Method

newRefinementLine(name, type,
                  start, end,
                  meshSize,
                  width)

Create refinement line of type "smooth" or "sharp" between start = [x,y,z] and end = [x,y,z] with a mesh size meshSize spread over a width width.

HOHQMesh.newSplineCurve — Method

newSplineCurve(name::String, nKnots::Int, data::Array{Float64,4})

Returns a spline curve given the number of knots and the array of knots.

HOHQMesh.newSplineCurve — Method

newSplineCurve(name::String, dataFile::String)

Returns a spline curve given a data file that contains the number of knots on the first line, and the spline data following that.

HOHQMesh.openProject — Method

openProject(fileName::String, folder::String)

Open existing project described in the control File.

folder   = folder the control file is in
fileName = the name of the file

HOHQMesh.plotProject! — Function

plotProject!(proj::Project, plotOptions::Int = 0)

Plot objects specified by the plotOptions. Construct the plotOptions by the sum of what is to be drawn from the choices MODEL, GRID, MESH, REFINEMENTS.

Example: To plot the model and the grid, plotOptions = MODEL + GRID. To plot just the mesh, plotOptions = MESH.

To plot everything, plotOptions = MODEL + GRID + MESH + REFINEMENTS

Contents are overlaid in the order: GRID, MESH, MODEL, REFINEMENTS

Requires Makie.jl

Please note that for this function to work, you need to load Makie.jl in your REPL (e.g., by calling using GLMakie).

HOHQMesh.postNotificationWithName — Method

postNotificationWithName(sender::Any, name::String, userInfo::Tuple)

Executes the function associated with the observer for the notification note

HOHQMesh.refinementRegionPoints — Method

refinementRegionPoints(r::Dict{String,Any})

Returns Array{Float64,2} being the plotting points of a refinement region

HOHQMesh.remove! — Method

remove!(proj::Project, curveName::String, innerBoundaryName::String)

Delete the curve named curveName from the inner boundary named innerBoundaryName

HOHQMesh.remove! — Method

remove!(proj::Project, curveName::String)

Delete the curve named curveName from the outer boundary

HOHQMesh.removeBackgroundGrid! — Method

removeBackgroundGrid!(proj::Project)

Remove the background grid block from the project.

HOHQMesh.removeInnerBoundary! — Method

removeInnerBoundary!(proj::Project, chainName::String)

Remove an entire inner boundary.

HOHQMesh.removeInnerBoundaryCurve! — Method

removeInnerBoundaryCurve!(proj::Project, name::String, chainName::String)

Remove the curve with name from an inner boundary chain with chainName.

HOHQMesh.removeInnerBoundaryCurveAtIndex! — Method

removeInnerBoundaryCurveAtIndex!(proj::Project, index::Int, chainName::String)

Remove the curve at index index from an inner boundary chain with chainName.

HOHQMesh.removeOuterBoundary! — Method

removeOuterBoundary!(proj::Project)

Remove the outer boundary curve if it exists.

HOHQMesh.removeOuterBoundaryCurveAtIndex! — Method

removeOuterBoundaryCurveAtIndex!(proj::Project, index::Int)

Remove a curve from the outer boundary chain at the specified index.

HOHQMesh.removeOuterBoundaryCurveWithName! — Method

removeOuterBoundaryCurveWithName!(proj::Project, name::String)

Remove the named curve in the outer boundary.

HOHQMesh.removeRefinementRegion! — Method

removeRefinementRegion!(proj::Project, name::String)

Delete the named refinement region.

HOHQMesh.removeRunParameters! — Method

removeRunParameters!(proj::Project)

Remove the run parameters block from the project. This is not undo-able.

HOHQMesh.removeSpringSmoother! — Method

removeSpringSmoother!(proj::Project)

Remove the background grid block from the project.

HOHQMesh.remove_mesh! — Method

remove_mesh!(proj::Project)

Remove the mesh file, control file, plot file, and stats file from proj.projectDirectory and delete the mesh from the plot.

HOHQMesh.renameCurve! — Method

renameCurve!(proj::Project, oldName::String, newName::String)

Any curve(s) on the outer boundary or in the inner boundary chain(s) with oldName are renamed with newName.

HOHQMesh.saveProject — Method

saveProject(proj::Project)

proj     = Project to be saved

Save a project dictionary to the file path specified when the project was created.

HOHQMesh.setArcCenter! — Method

setArcCenter!(crv::Dict{String,Any}, point::Array{Float64})

Set the center of a circular or elliptic arc.

HOHQMesh.setArcEndAngle! — Method

setArcEndAngle!(arc::Dict{String,Any}, angle::Float64)

Sets a new value for the end angle of a CIRCULAR_ARC or ELLIPTIC_ARC.

HOHQMesh.setArcRadius! — Function

setArcRadius!(arc::Dict{String,Any}, radius::Float64, key::String)

Sets a new value for the radius of an arc. Default behavior expects a CIRCULAR_ARC with key = "radius". For an ELLIPTIC_ARC use either key = "xRadius" or key = "yRadius".

HOHQMesh.setArcRotation! — Method

setArcRotation!(arc::Dict{String,Any}, rotation::Float64)

Sets a new rotation angle of an ELLIPTIC_ARC.

HOHQMesh.setArcStartAngle! — Method

setArcStartAngle!(arc::Dict{String,Any}, angle::Float64)

Sets a new value for the start angle of a CIRCULAR_ARC or ELLIPTIC_ARC.

HOHQMesh.setArcUnits! — Method

setArcUnits!(crv::Dict{String,Any}, units::String)

Set the units for the start and end angles of a circular or elliptic arc curve.

HOHQMesh.setBackgroundGridLowerLeft! — Method

setBackgroundGridLowerLeft!(proj::Project, x0::Array{Float64})

Set the lower left location of the background grid for problems that have no outer boundary.

HOHQMesh.setBackgroundGridSize! — Function

setBackgroundGridSpacing!(proj::Project, dx::Float64, dy::Float64, dz::Float64 = 0.0)

User facing function

HOHQMesh.setBackgroundGridSize! — Method

setBackgroundGridSize!(proj::Project, dx::Array{Float64}, key::String)

Set the grid size dx of an existing background grid within proj. Here, dx is passed an array.

HOHQMesh.setBackgroundGridSize! — Method

setBackgroundGridSize!(proj::Project, dx::Float64, dy::Float64, key::String)

Set the grid size dx of an existing background grid within proj. Here, the new grid size in either direction is passed individually with dxand dy.

HOHQMesh.setBackgroundGridSteps! — Method

setBackgroundGridSteps!(proj::Project, N::Array{Int})

Set how many steps of size setBackgroundGridSpacing in each direction the background grid extends from the lower left.

HOHQMesh.setCurveName! — Method

setCurveName!(curveDict, name)

Set the name of the curve represented by curveDict.

HOHQMesh.setEndPoint! — Method

setEndPoint!(crv::Dict{String,Any}, point::Array{Float64})

Set the end point for a line curve.

HOHQMesh.setInnerBoundaryChainConnect! — Method

setInnerBoundaryChainConnect!(proj::Project, chainName::String, connect::String)

Sets the connect setting for the boundary curve optimization of an inner boundary CHAIN with chainName.

By default, boundary curve optimization is done curve by curve within the chain with an "empty" connect = "[]" key, as usually there will be discontinuities (e.g. corners) between the curves. If it is desired to define a more global approximation across multiple curves update the connect key. For example, if the chain has four segments then connect = "2-3" requests that optimization occurs across curve segments two and three in the list, where as curves 1 and 4 are optimized individually.

HOHQMesh.setInnerBoundaryChainContinuity! — Method

setInnerBoundaryChainContinuity!(proj::Project, chainName::String, continuity::Int)

Sets the continuity (number of continuous derivatives) for the boundary curve optimization of an inner boundary CHAIN with chainName.

HOHQMesh.setInnerBoundaryChainOptimizeStatus! — Method

setInnerBoundaryChainOptimizeStatus!(proj:Project, chainName::String, type::String)

Adjust the type for the boundary curve optimization of an inner boundary CHAIN with chainName. Available types are none, L2Norm, or H1Norm.

HOHQMesh.setInnerBoundaryChainTolerance! — Method

setInnerBoundaryChainTolerance!(proj::Project, chainName::String, tolerance::Float64)

Sets the tolerance for the boundary curve optimization of an inner boundary CHAIN with chainName.

HOHQMesh.setMeshFileFormat! — Method

setMeshFileFormat!(proj::Project, meshFileFormat::String)

Set the file format for the mesh file. Acceptable choices are "ISM", "ISM-V2", or "ABAQUS".

HOHQMesh.setName! — Method

setName!(proj::Project,name::String)

The name of the project is the filename to be used by the mesh, plot, and stats files. It is also the name of the control file the tool will produce.

HOHQMesh.setOuterBoundaryConnect! — Method

setOuterBoundaryConnect!(proj::Project, connect::String)

Sets the connect setting for the boundary optimization of the outer boundary chain.

By default, boundary curve optimization is done curve by curve within the chain with an "empty" connect = "[]" key, as usually there will be discontinuities (e.g. corners) between the curves. If it is desired to define a more global approximation across multiple curves update the connect key. For example, if the chain has four segments then connect = "2-3" requests that optimization occurs across curve segments two and three in the list, where as curves 1 and 4 are optimized individually.

HOHQMesh.setOuterBoundaryContinuity! — Method

setOuterBoundaryContinuity!(proj::Project, continuity::Int)

Sets the continuity (number of continuous derivatives) for the boundary optimization of the outer boundary chain.

HOHQMesh.setOuterBoundaryOptimizeStatus! — Method

setOuterBoundaryOptimizeStatus!(proj:Project, type::String)

Adjust the type for the boundary curve optimization for the outer boundary chain. Available types are none, L2Norm, or H1Norm.

HOHQMesh.setOuterBoundaryTolerance! — Method

setOuterBoundaryTolerance!(proj::Project, tolerance::Float64)

Sets the tolerance for the boundary optimization of the outer boundary chain.

HOHQMesh.setPlotFileFormat! — Method

setPlotFileFormat!(proj::Project, plotFileFormat::String)

Set the file format for the plot file. Acceptable choices are "sem", which includes interior nodes and boundary nodes and "skeleton", which includes only the corner nodes.

HOHQMesh.setPolynomialOrder! — Method

setPolynomialOrder!(proj::Project, p::Int)

Set the polynomial order for boundary curves in the mesh file to p.

HOHQMesh.setRefinementEnd! — Method

setRefinementEnd!(refinementRegion, location)

Set the end point location of a refinement line, location = [x, y, z].

HOHQMesh.setRefinementGridSize! — Method

setRefinementGridSize!(r::Dict{String,Any}, h)

Set the grid size, h for the refinement region. r is the dictionary that represents the refinement region.

HOHQMesh.setRefinementLocation! — Method

setRefinementLocation!(refinementCenter, location)

Set the location of a refinement center to location = [x,y,z].

HOHQMesh.setRefinementName! — Method

setRefinementName!(r::Dict{String,Any}, type)

Set a name for the refinement region.r is the dictionary that represents the refinement region.

HOHQMesh.setRefinementStart! — Method

setRefinementStart!(refinementRegion, location)

Set the start point location of a refinement line, location = [x, y, z].

HOHQMesh.setRefinementType! — Method

setRefinementType!(refinementRegion, type)

Set the type, either "smooth" or "sharp" for the given refinement region.

HOHQMesh.setRefinementWidth! — Method

setRefinementWidth!(r::Dict{String,Any}, width)

Set the width of the refinement region. r is the dictionary that represents the refinement region.

HOHQMesh.setSmoothingIterations! — Method

setSmoothingIterations!(proj::Project, iterations::Int)

Set the number of iterations to smooth the mesh.

HOHQMesh.setSmoothingStatus! — Method

setSmoothingStatus!(proj:Project, status::String)

Status is either "ON" or "OFF"

HOHQMesh.setSmoothingType! — Method

setSmoothingType!(proj:Project, type::String)

Type is either LinearSpring or LinearAndCrossbarSpring

HOHQMesh.setSplineNKnots! — Method

setSplineNKnots!(spline::Dict{String,Any}, nKnots::Int)

Set a new value for the number of points (or knots) in a spline.

HOHQMesh.setSplinePoints! — Method

setSplinePoints!(spline::Dict{String,Any},points::Array{Float64,4})

Set the parametric spline point data into the spline curve dictionary.

HOHQMesh.setStartPoint! — Method

setStartPoint!(crv::Dict{String,Any}, point::Array{Float64})

Set the start point for a line curve.

HOHQMesh.setXEqn! — Method

setXEqn!(parametricEquationCurve, eqn)

For a parametric equation, set the x-equation.

HOHQMesh.setYEqn! — Method

setYEqn!(parametricEquationCurve, eqn)

For a parametric equation, set the y-equation.

HOHQMesh.setZEqn! — Method

setZEqn!(parametricEquationCurve, eqn)

For a parametric equation, set the zEqn-equation.

HOHQMesh.unRegisterForNotification — Method

unRegisterForNotification(observer::Any, note::String)

Remove the observer from being notified by the notification note

HOHQMesh.updatePlot! — Function

updatePlot!(proj::Project, plotOptions::Int)

Replot with the new plotOptions = combinations (sums) of

GRID, MESH, MODEL, REFINEMENTS

Example: updatePlot!(p, MESH + MODEL)

Requires Makie.jl

Please note that for this function to work, you need to load Makie.jl in your REPL (e.g., by calling using GLMakie).

generate_mesh(control_file; output_directory="out", mesh_filename=nothing, plot_filename=nothing, stats_filename=nothing, verbose=false, subdivision_maximum=8)
Generate a mesh based on the control_file with the HOHQMesh mesh generator and store resulting files in output_directory.