HOHQMesh.jl API

HOHQMesh.add! — Method

add!(proj::Project, crv::Dict{String,Any}, boundaryName::String)

Add a curve to the inner boundary named boundaryName.

HOHQMesh.add! — Method

add!(proj::Project, obj::Dict{String,Any})

Add a curve to the outer boundary or a refinement reion to the project

HOHQMesh.addBackgroundGrid! — Method

addBackgroundGrid!(proj::Project, x0::Array{Float64}, dx::Array{Float64}, N::Array{Int})

Add the background grid block using the left corner, x0, the grid size dx, and the number of intervals in each direction. Use this when there is no outer boundary defined in the model. This version mimics HOHQMesh's backgroundGrid block, but the version

addBackgroundGrid!(proj::Project, box::Array{Float64},  N::Array{Int} )

is a lot easier to use.

TODO: Change HOHQMesh and delete this way to specify the domain and use the bounding box one instead.

HOHQMesh.addBackgroundGrid! — Method

addBackgroundGrid!(proj::Project, box::Array{Float64},  N::Array{Int} )

Add the background grid block with bounding box = [TOP, LEFT, BOTTOM, RIGHT] and the number of intervals in each diredction. Use this when there is no outer boundary defined in the model.

HOHQMesh.addBackgroundGrid! — Method

addBackgroundGrid(proj::Project, bgSize::Array{Float64})

Add the background grid block with the grid size to be a 3-vector. Use this when there is an outer boundary defined in the model.

HOHQMesh.addBackgroundGrid! — Method

addBackgroundGrid!(proj::Project, dict::Dict{String,Any})

Used only for undo/redo.

HOHQMesh.addCurveToInnerBoundary! — Method

addCurveToInnerBoundary!(proj::Project, crv::Dict{String,Any}, boundaryName::String)

Add a curve to the inner boundary with name boundaryName. If an inner boundary of that name does not exist, one is created.

HOHQMesh.addCurveToOuterBoundary! — Method

addCurveToOuterBoundary!(proj::Project, crv::Dict{String,Any})

Add a curve to the outer boundary. The curves must be added in order counter-clockwise

HOHQMesh.addInnerBoundaryWithName! — Method

addInnerBoundaryWithName!(proj::Project,name::String)

Create a new empty inner boundary with the given name.

HOHQMesh.addObserver — Method

addObserver(observer::Any, note::String, fnction::Any)

fnction is the function to be executed (called) when a notification of name note is given.

The function called upon notification must have the signature fnction(observer, sender, args...)

HOHQMesh.addOuterBoundary! — Method

addOuterBoundary!(proj::Project)

Add an empty outer boundary to the project. There can be only one.

HOHQMesh.addRefinementRegion! — Method

addRefinementRegion!(proj::Project,r::Dict{String,Any})

Add the refinement region to the project

HOHQMesh.addRefinementRegionPoints! — Method

addRefinementRegionPoints!(proj::Project, r::Dict{String,Any})

Compute and add to the project the plotting points for the refinement region

HOHQMesh.addRunParameters! — Function

addRunParameters!(proj::Project, 
    plotFormat::String     = "skeleton", 
    meshFileFormat::String = "ISM-V2", 
    polynomialOrder::Int   = 5)

Add a RUN_PARAMETERS block and set all the parameters in one call.

HOHQMesh.addSpringSmoother! — Function

addSpringSmoother!(status::String, type::String, nIterations::Int)

status is either `ON` or `OFF`
type   is either `LinearSpring` or `LinearAndCrossbarSpring`

HOHQMesh.bboxUnion — Method

bboxUnion(box1::Array{Float64}, box2::Array{Float64})

Returns the union of two bounding boxes

HOHQMesh.curveBounds — Method

curveBounds(crvPoints::Array{Float64,2})

Find the bounds of a single curve, discretized as an array

HOHQMesh.duplicateCurve — Method

duplicateCurve(crv::Dict{String,Any}, newName::String)

Duplicate the given curve giving it the new name.

HOHQMesh.emptyBounds — Method

Returns an array that will always be ignored when unioned with another bounding box.

HOHQMesh.examples_dir — Method

examples_dir()

Return the path to the directory with some example mesh setups.

HOHQMesh.generate_mesh — Method

generate_mesh(control_file;
              output_directory="out",
              mesh_filename=nothing, plot_filename=nothing, stats_filename=nothing)

Generate a mesh based on the control_file with the HOHQMesh mesh generator and store resulting files in output_directory.

You can set the mesh filename, the plot filename, and the statistics filename using the keyword arguments mesh_filename, plot_filename, and stats_filename, respectively. If set to nothing, the filenames for the mesh file, plot file, and statistics file are generated automatically from the control file name. For example, path/to/ControlFile.control will result in output files ControlFile.mesh, ControlFile.tec, and ControlFile.txt.

You can activate verbose output from HOHQMesh that prints additional messages and debugging mesh information with the keyword argument verbose.

This function returns the output to stdout of the HOHQMesh binary when generating the mesh.

HOHQMesh.get — Method

get(proj::Project, curveName::String, boundaryName::String)

Get the curve named curveName from the inner boundary named boundaryName

HOHQMesh.get — Method

get(proj::Project, curveName::String)

Get the curve with name curveName from the outer boundary.

HOHQMesh.getAllRefinementRegions — Method

allRefinementRegions(proj::Project)

Get the list of refinement regions.

HOHQMesh.getArcCenter — Method

getArcCenter(crv::Dict{String,Any}, point::Array{Float64})

Get the center of a circular arc as an array

HOHQMesh.getArcEndAngle — Method

getArcEndAngle(arc::Dict{String,Any}, angle::Float64)

HOHQMesh.getArcRadius — Method

getArcRadius(arc::Dict{String,Any}, radius::Float64)

HOHQMesh.getArcStartAngle — Method

getArcStartAngle(arc::Dict{String,Any}, angle::Float64)

HOHQMesh.getArcUnits! — Method

getArcUnits(crv::Dict{String,Any}, units::String)

Get the units for the start and end angles of a circular arc curve.

HOHQMesh.getBackgroundGridLowerLeft — Method

function getBackgroundGridLowerLeft(proj::Project)

Returns the [x,y] of the lower left point of thebackground grid.

HOHQMesh.getBackgroundGridSize — Method

getBackgroundGridSize(proj::Project)

Returns the background grid size array.

HOHQMesh.getBackgroundGridSteps — Method

function getBackgroundGridLowerLeft(proj::Project)

Returns the [x,y,z] of the lower left point of thebackground grid.

HOHQMesh.getCurveName — Method

getCurveName(crv::Dict{String,Any})

HOHQMesh.getCurveType — Method

getCurveType(crv::Dic{String,Any})

Get the type of the curve, `END_POINTSLINE_CURVE`, `PARAMETRIC_EQUATION_CURVE`,
`SPLINE_CURVE`, or `CIRCULAR_ARC` as a string.

HOHQMesh.getEndPoint — Method

getEndPoint(crv::Dict{String,Any}, point::Array{Float64})

Get the end point for a line curve as an array.

HOHQMesh.getInnerBoundary — Method

getInnerBoundary(proj::Project, name::String)

Get the chain of curves from the inner boundary with name name.

HOHQMesh.getInnerBoundaryChainWithName — Method

getInnerBoundaryWithName(proj::Project, name::String)

Get the inner boundary CHAIN with the given name. If one does not exist, it is created.

HOHQMesh.getInnerBoundaryCurve — Method

HOHQMesh.getMeshFileFormat — Method

getMeshFileFormat(proj::Project)

Returns the format in which the mesh will be written.

HOHQMesh.getName — Method

getName(proj::Project)

Returns the filename to be used by the mesh, plot, control, and stats files.

HOHQMesh.getOuterBoundaryChainList — Method

getOuterBoundary(proj::Project)

Get the array of outer boundary curves.

HOHQMesh.getOuterBoundaryCurveWithName — Method

getOuterBoundaryCurveWithName(proj::Project, name::String)

HOHQMesh.getPlotFileFormat — Method

getPlotFileFormat(proj::Project)

Returns the plot file format.

HOHQMesh.getPolynomialOrder — Method

getPolynomialOrder(proj::Project)

Returns the polynomial order for boundary curves in the mesh file.

HOHQMesh.getRefinementEnd — Method

getRefinementEnd(r::Dict{String,Any})

Return Array{Float64} of the end location of the refinement line

HOHQMesh.getRefinementGridSize — Method

getRefinementGridSize(r::Dict{String,Any})

Returns the grid size,h, as Float64. r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementLocation — Method

getRefinementLocation(r::Dict{String,Any})

Return Array{Float64} of the location of the refinement center.r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementName — Method

getRefinementName(r::Dict{String,Any})

Return name of the refinement. r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementRegion — Method

getRefinementRegion(proj::Project, indx)

Get the refinement region with index, indx from the project. Returns nothing if there is none. The return value is a dictionary that represents the refinement region.

HOHQMesh.getRefinementRegion — Method

(i,r) = getRefinementRegion(project, name)

Get the refinement region with the given name and its location in the list of refinement regions.

HOHQMesh.getRefinementStart — Method

getRefinementStart  (r::Dict{String,Any})

Return Array{Float64} of the start location of the refinement line. r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementType — Method

getRefinementType(r::Dict{String,Any})

Return the type of refinement, either "smooth" or "sharp". r is the dictionary that represents the refinement region.

HOHQMesh.getRefinementWidth — Method

getRefinementWidth(r::Dict{String,Any})

Returns the region width,w, as Float64. r is the dictionary that represents the refinement region.

HOHQMesh.getSmoothingIterations — Method

HOHQMesh.getSmoothingStatus — Method

smoothingStatus(proj::Project)

Returns whether the smoother will be "ON" or "OFF"

HOHQMesh.getSmoothingType — Method

smoothingType(proj::Project)

Returns either "LinearSpring" or "LinearAndCrossbarSpring"

HOHQMesh.getSplineNKnots — Method

getSplineNKnots(spline::Dict{String,Any})

HOHQMesh.getSplinePoints — Method

getSplinePoints(spline::Dict{String,Any})

HOHQMesh.getStartPoint — Method

getStartPoint(crv::Dict{String,Any}, point::Array{Float64})

Get the start point for a line curve as an array

HOHQMesh.getXEqn — Method

getXEqn(crv::Dict{String,Any})

HOHQMesh.getYEqn — Method

getYEqn(crv::Dict{String,Any})

HOHQMesh.getZEqn — Method

getZEqn(crv::Dict{String,Any})

HOHQMesh.getfolder — Method

path(proj::Project)

Returns the directory where the project files will be written

HOHQMesh.hasBackgroundGrid — Method

hasBackgroundGrid(proj::Project)

Tests to see if the project has a backgroundGrid dictionary defined.

HOHQMesh.innerBoundaryIndices — Method

innerBoundaryIndices(proj::Project, curveName::String)

Returns (curveIndex,chainIndex) for the location of the curve named curveName in it's inner boundary chain.

HOHQMesh.insertOuterBoundaryCurveAtIndex! — Method

insertOuterBoundaryCurveAtIndex(proj::Project, crv::Dict{String,Any}, indx::Int)

Insert a curve at the specified index.

HOHQMesh.insertRefinementRegion! — Method

insertRefinementRegion!(proj::Project, r::Dict{String,Any}, indx::Int)

Used by undo()

HOHQMesh.new — Function

new(name::String, 
    center::Array{Float64},  
    radius::Float64,
    startAngle::Float64, 
    endAngle::Float64,
    units::String)

Create a new circular arc.

HOHQMesh.new — Function

new(name::String, 
    xEqn::String, 
    yEqn::String, 
    zEqn::String = "z(t) = 0.0" )

Create a new parametric equation curve.

HOHQMesh.new — Method

new(name::String,
         xStart::Array{Float64},
         xEnd::Array{Float64})

Create a new line defined by its end points.

HOHQMesh.new — Method

new(name::String, nKnots::Int, data::Matrix{Float64})

Create a spline curve from an array of knots

HOHQMesh.new — Method

new(name::String, dataFile::String)

Create a spline curve from the contents of a data file.

HOHQMesh.newCircularArcCurve — Function

newCircularArcCurve(name::String, center::Array{Float64},  
    startAngle::Float64, endAngle::Float64,
    units::String)

Creates and returns a new circular arc curve in the form of a Dictionary

HOHQMesh.newEndPointsLineCurve — Method

newEndPointsLineCurve(name::String, xStart::Array{Float64},xEnd::Array[Float64])

Creates and returns a new curve defined by its end points in the form of a Dictionary

HOHQMesh.newParametricEquationCurve — Function

newParametricEquationCurve(name::String, 
    xEqn::String, 
    yEqn::String, 
    zEqn::String = "z(t) = 0.0" )

Creates and returns a new parametricEquationCurve in the form of a Dictionary

HOHQMesh.newProject — Method

newProject(name::String, folder::String)

Create a new project with the given name. That name will be used for the mesh and plot files in the specified folder.

HOHQMesh.newRefinementCenter — Method

newRefinementCenter(type, center, meshSize, width )

Create refinement center of type "smooth" or "sharp" centered at center = [x,y,z]with a mesh sizemeshSizespread over a radiuswidth`.

HOHQMesh.newRefinementLine — Method

newRefinementLine(type, 
                    start, end,
                    meshSize,
                    width )

Create refinement line of type "smooth" or "sharp" between start = [x,y,z] and end = [x,y,z] with a mesh size meshSize spread over a width width.

HOHQMesh.newSplineCurve — Method

newSplineCurve(name::String, nKnots::Int, data::Array{Float64,4})

Returns a spline curve given the number of knots and the array of knots.

HOHQMesh.newSplineCurve — Method

newSplineCurve(name::String, dataFile::String)

Returns a spline curve given a data file that contains the number of knots on the first line, and the spline data following that.

HOHQMesh.openProject — Method

openProject(fileName::String, folder::String)

Open existing project described in the control File.

folder   = folder the control file is in
fileNmae = the name of the file

HOHQMesh.plotProject! — Function

plotProject!(proj::Project, plotOptions::Int = 0)

Plot objects specified by the plotOptions. Construct the plotOptions by the sum of what is to be drawn from the choices MODEL, GRID, MESH, REFINEMENTS.

Example: To plot the model and the grid, plotOptions = MODEL + GRID. To plot just the mesh, plotOptions = MESH.

To plot everything, plotOptions = MODEL + GRID + MESH + REFINEMENTS

Contents are overlayed in the order: GRID, MESH, MODEL, REFINEMENTS

HOHQMesh.postNotificationWithName — Method

postNotificationWithName(sender::Any, name::String, userInfo::Tuple)

Executes the function associated with the observer for the notification note

HOHQMesh.refinementRegionCenter — Method

refinementRegionCenter(r::Dict{String,Any})

Get, or compute, the center of the given refinement region.

HOHQMesh.refinementRegionPoints — Method

refinementRegionPoints(r::Dict{String,Any})

Returns Array{Float64,2} being the plotting points of a refinement region

HOHQMesh.remove! — Method

remove!(proj::Project, curveName::String, innerBoundaryName::String)

Delete the curve named curveName from the inner boundary named innerBoundaryName

HOHQMesh.remove! — Method

remove!(proj::Project, curveName::String)

Delete the curve named curveName from the outer boundary

HOHQMesh.removeBackgroundGrid! — Method

removeBackgroundGrid!(proj::Project)

Remove the background grid block from the project.

HOHQMesh.removeInnerBoundary! — Method

removeInnerBoundary!(proj::Project, chainName::String)

Remove an entire inner boundary

HOHQMesh.removeInnerBoundaryCurve! — Method

removeInnerBoundaryCurve!(proj::Project, name::String)

Remove the named curve in the outer boundary

HOHQMesh.removeOuterBoundaryCurveWithName! — Method

removeOuterBoundaryCurveWithName!(proj::Project, name::String)

Remove the named curve in the outer boundary

HOHQMesh.removeOuterboundary! — Method

removeOuterboundary!(proj::Project)

Remove the outer boundary curve if it exists.

HOHQMesh.removeRefinementRegion! — Method

removeRefinementRegion!(proj::Project, name::String)

Delete the named refinement region.

HOHQMesh.removeRunParameters! — Method

removeRunParameters!(proj::Project)

Remove the run parameters block from the project.

HOHQMesh.removeSpringSmoother! — Method

removeSpringSmoother!(proj::Project)

Remove the background grid block from the project.

HOHQMesh.saveProject — Method

saveProject(proj::Project)

proj     = Project to be saved

Save a project dictionary to the file path specified when the project was created.

HOHQMesh.setArcCenter! — Method

setArcCenter!(crv::Dict{String,Any}, point::Array{Float64})

Set the center of a circular arc.

HOHQMesh.setArcEndAngle! — Method

setArcEndAngle!(arc::Dict{String,Any}, angle::Float64)

HOHQMesh.setArcRadius! — Method

setArcRadius!(arc::Dict{String,Any}, radius::Float64)

HOHQMesh.setArcStartAngle! — Method

setArcStartAngle!(arc::Dict{String,Any}, angle::Float64)

HOHQMesh.setArcUnits! — Method

setArcUnits(crv::Dict{String,Any}, units::String)

Set the units for the start and end angles of a circular arc curve.

HOHQMesh.setBackgroundGridLowerLeft! — Method

setBackgroundGridLowerLeft!(proj::Project, x0::Array{Float64})

Set the lower left location of the background grid for problems that have no outer boundary.

HOHQMesh.setBackgroundGridSize! — Function

setBackgroundGridSpacing!(proj::Project, dx::Float64, dy::Float64, dz::Float64 = 0.0)

User facing function

HOHQMesh.setBackgroundGridSize! — Method

setBackgroundGridSize!(proj::Project, dx::Array{Float64},key::String)

HOHQMesh.setBackgroundGridSize! — Method

setBackgroundGridSize!(proj::Project, dx::Float64, dy::Float64,key::String)

HOHQMesh.setBackgroundGridSteps! — Method

setBackgroundGridSteps!(proj::Project, N::Array{Int})

Set how many steps of size setBackgroundGridSpacing in each direction the background grid extends from the lower left.

HOHQMesh.setCurveName! — Method

setCurveName!(curveDict, name)

Set the name of the curve represented by curveDict.

HOHQMesh.setEndPoint! — Method

setEndPoint!(crv::Dict{String,Any}, point::Array{Float64})

Set the end point for a line curve.

HOHQMesh.setFolder — Method

setFolder(proj::Project,folder::String)

Set the path to the directory where the mesh, plot, control, and stats files will be written

HOHQMesh.setMeshFileFormat! — Method

setMeshFileFormat(proj::Project, meshFileFormat::String)

Set the file format for the mesh file. Acceptable choices are "ISM" and "ISM-V2".

HOHQMesh.setName! — Method

setName(proj::Project,name::String)

The name of the project is the filename to be used by the mesh, plot, and stats files. It is also the name of the control file the tool will produce.

HOHQMesh.setPlotFileFormat! — Method

setPlotFileFormat(proj::Project, plotFileFormat::String)

Set the file format for the plot file. Acceptable choices are "sem", which includes interior nodes and boundary nodes and "skeleton", which includes only the corner nodes.

HOHQMesh.setPolynomialOrder! — Method

setPolynomialOrder(proj::Project, p::Int)

Set the polynomial order for boundary curves in the mesh file to p.

HOHQMesh.setRefinementEnd! — Method

setRefinementEnd(refinementRegion, location)

Set the end point location of a refinement line, location = [x, y, z].

HOHQMesh.setRefinementGridSize! — Method

setRefinementGridSize(r::Dict{String,Any}, h)

Set the grid size, h for the refinement region. r is the dictionary that represents the refinement region.

HOHQMesh.setRefinementLocation! — Method

setRefinementLocation!(refinementCenter, location)

Set the location of a refinement center to location = [x,y,z].

HOHQMesh.setRefinementName! — Method

setRefinementName!(r::Dict{String,Any}, type)

Set a name for the refinement region.r is the dictionary that represents the refinement region.

HOHQMesh.setRefinementStart! — Method

setRefinementStart!(refinementRegion, location)

Set the start point location of a refinement line, location = [x, y, z].

HOHQMesh.setRefinementType! — Method

setRefinementType!(refinementRegion, type)

Set the type, either "smooth" or "sharp" for the given refinement region.

HOHQMesh.setRefinementWidth! — Method

setRefinementWidth!(r::Dict{String,Any}, width)

Set the width of the refinement region. r is the dictionary that represents the refinement region.

HOHQMesh.setSmoothingIterations! — Method

setSmoothingIterations!((proj::Project, iterations::Int)

Set the number of iterations to smooth the mesh.

HOHQMesh.setSmoothingStatus! — Method

setSmoothingStatus(proj:Project, status::String)

status is either "ON" or "OFF"

HOHQMesh.setSmoothingType! — Method

setSmoothingType!(proj:Project, status::String)

type is either LinearSpring or LinearAndCrossbarSpring

HOHQMesh.setSplineNKnots! — Method

setSplineNKnots!(spline::Dict{String,Any}, nKnots::Int)

HOHQMesh.setSplinePoints! — Method

setSplinePoints!(spline::Dict{String,Any},points::Array{Float64,4})

HOHQMesh.setStartPoint! — Method

setStartPoint!(crv::Dict{String,Any}, point::Array{Float64})

Set the start point for a line curve.

HOHQMesh.setXEqn! — Method

setXEqn!(parametricEquationCurve, eqn)

For a parametric equation, set the x-equation.

HOHQMesh.setYEqn! — Method

setYEqn!(parametricEquationCurve, eqn)

For a parametric equation, set the y-equation.

HOHQMesh.setZEqn! — Method

setZEqn!(parametricEquationCurve, eqn)

For a parametric equation, set the zEqn-equation.

HOHQMesh.unRegisterForNotification — Method

unRegisterForNotification(observer::Any, note::String)

Remove the observer from being notified by the notification note

HOHQMesh.updatePlot! — Method

updatePlot!(proj::Project, plotOptions::Int)

Replot with the new plotOptions = combinations (sums) of

GRID, MESH, MODEL, REFINEMENTS

Example: updatePlot(p, MESH + MODEL)

HOHQMesh.updatePlot! — Method

updatePlot!(proj::Project)

This version replots the figure with the current options. Legacy.